Improve robustness of Kokkos build script
It seems that, sometimes, the kokkos CMake will change the CMAKE_CXX_COMPILER
to an absolute path before storing it in the cache. Subsequent calls to cmake
that do not have -DCMAKE_CXX_COMPILER=$absolute_path_to_cxx will be interpreted
as a changed compiler, which results in the CMakeCache being flushed and then
results in CMAKE_INSTALL_PREFIX being lost.

I'm not sure if this is a bug in CMake or Kokkos' cmake files, but this PR will
work around the problem.

[BFB]